+++ src/index.ts
@@ -1,9 +1,9 @@
 export * from './types';
+export * from './route';
 export * from './server';
 export * from './client';
 export * from './constants';
-export * from './controllers';
 export * from './routes/files';
-export * from './plugins/request';
+export * from './plugins/access';
 export * from './plugins/response';
 export * as cookies from './plugins/cookies';