+++ src/route.ts
@@ -1,38 +1,38 @@
 import { constants } from 'http2';
 
-import { Handler } from './types';
+import { RequestHandler } from './types';
 import { addBounds, normalize } from './utils';
 
 export interface Route {
   readonly path: string;
   readonly method: string;
-  handle: Handler;
+  handle: RequestHandler;
 }
 
 export const route = (
   method: string,
   path: string,
-  handler: Handler
+  handler: RequestHandler
 ): Route => ({
   method,
   path: addBounds(path),
   handle: handler,
 });
 
-export const del = (path: string, handler: Handler): Route =>
+export const del = (path: string, handler: RequestHandler): Route =>
   route(constants.HTTP2_METHOD_DELETE, path, handler);
-export const put = (path: string, handler: Handler): Route =>
+export const put = (path: string, handler: RequestHandler): Route =>
   route(constants.HTTP2_METHOD_PUT, path, handler);
-export const get = (path: string, handler: Handler): Route =>
+export const get = (path: string, handler: RequestHandler): Route =>
   route(constants.HTTP2_METHOD_GET, path, handler);
-export const head = (path: string, handler: Handler): Route =>
+export const head = (path: string, handler: RequestHandler): Route =>
   route(constants.HTTP2_METHOD_HEAD, path, handler);
-export const post = (path: string, handler: Handler): Route =>
+export const post = (path: string, handler: RequestHandler): Route =>
   route(constants.HTTP2_METHOD_POST, path, handler);
-export const options = (path: string, handler: Handler): Route =>
+export const options = (path: string, handler: RequestHandler): Route =>
   route(constants.HTTP2_METHOD_OPTIONS, path, handler);
 
-export const controller =
+export const group =
   (prefix: string = '') =>
   (...locals: ReadonlyArray<Route>): ReadonlyArray<Route> =>
     locals.map(({ method, path, handle }) => ({